Update hud_obj_sapper.res
Fixed the building progress panel being misaligned
I'm mentioning this here because this is only ever on screen for at most 0.1 seconds at a time so you're likely not even aware this was a problem in the first place. This is so minor in fact that I won't even mention it when I push the next update, whenever that may be.